feat: Move pin-related code to SWM
